Fix tests and move classes out of sync_apps.py
